// src/app/models/member.ts
export class Member {
    id: string;
    memberRef: string;
    password: string;
    memberName: string;
    firstName: string;
    lastName: string;
    address: string;
    email: string;
    telephone: string;
    status: string;
    active: string;
    verify: boolean;
    createDate: Date;
    createBy: string;
    approve1Date: Date;
    approve1By: string;
    approve2Date: Date;
    approve2By: string;
    imageBookBank: string;
    imagePassport: string;
    imageBookBankOri: string;
    imagePassportOri: string;

    constructor() {
    }
}